refatora routes
